+++ src/app/(auth)/logout/route.ts
@@ -1,0 +1,9 @@
+import { type NextRequest, NextResponse } from "next/server";
+
+import { getSupabaseServerClient } from "@/lib/supabase/server";
+
+export async function GET(request: NextRequest) {
+  const supabase = await getSupabaseServerClient();
+  await supabase.auth.signOut();
+  return NextResponse.redirect(new URL("/", request.url));
+}
