+++ src/app/surveys/page.tsx
@@ -66,7 +66,7 @@ export default function SurveyPage() {
       </div>
 
       {/* 시작하기 버튼 */}
-      <div className="fixed bottom-0 mb-2 w-full max-w-lg p-5">
+      <div className="fixed bottom-0 mb-2 w-full max-w-md p-5">
         {/* TODO: 제대로 수정 */}
         <Link href="/">
           <Button className="h-12 w-full gap-2 bg-[#0056EB] px-3 py-2 text-lg font-medium text-white hover:bg-blue-700">
