+++ src/assets/kakao-logo.tsx
@@ -1,0 +1,14 @@
+export const KakaoLogo = () => {
+  return (
+    <svg width="18" height="19" viewBox="0 0 18 19" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <g clipPath="url(#clip0_15680_2954)">
+        <path fillRule="evenodd" clipRule="evenodd" d="M9.00002 1.31357C4.02917 1.31357 0 4.42653 0 8.26585C0 10.6536 1.5584 12.7585 3.93152 14.0105L2.93303 17.6581C2.84481 17.9804 3.21341 18.2373 3.49646 18.0505L7.87334 15.1618C8.2427 15.1974 8.61808 15.2182 9.00002 15.2182C13.9705 15.2182 17.9999 12.1054 17.9999 8.26585C17.9999 4.42653 13.9705 1.31357 9.00002 1.31357Z" fill="black" />
+      </g>
+      <defs>
+        <clipPath id="clip0_15680_2954">
+          <rect width="17.9999" height="18" fill="white" transform="translate(0 0.713562)" />
+        </clipPath>
+      </defs>
+    </svg>
+  );
+};
