+++ src/features/homePage/components/BottomButtonGroup.tsx
@@ -8,7 +8,7 @@ interface IBottomButtonGroup {
 
 export default function BottomButtonGroup({ bottomButtonGroup }: IBottomButtonGroup) {
   return (
-    <section className="fixed bottom-0 w-full max-w-lg rounded-t-xl bg-[#E5E7EB] p-4 pb-5">
+    <section className="fixed bottom-0 w-full max-w-md rounded-t-xl bg-[#E5E7EB] p-4 pb-5">
       <div className="flex gap-2">
         {bottomButtonGroup.map(item => (
           <button key={item.id} className="flex w-full gap-4 rounded-lg bg-white px-5 py-6 font-semibold text-black">
